rename lib/fly to lib/api
