+++ lib/api/-clear.js
