+++ lib/api/-parallel.js
