+++ lib/api/-serial.js
