+++ lib/cli/spawn.js
@@ -3,7 +3,7 @@
 const Promise = require('bluebird');
 const find = require('../utils/find');
 const loadPlugins = require('../plugins').load
-const Fly = require('../fly');
+const Fly = require('../api');
 const co = Promise.coroutine;
 
 /**
